refactor: move buildContractBasedSettlement to TransactionBatchGraph
