+++ src/apps/explorer/components/TransanctionBatchGraph/utils.ts
@@ -5,7 +5,7 @@ import { Account, ALIAS_TRADER_NAME, Transfer } from 'api/tenderly'
 import { TypeEdgeOnTx, TypeNodeOnTx } from 'apps/explorer/components/TransanctionBatchGraph/types'
 import { OrderKind } from '@cowprotocol/cow-sdk'
 import { abbreviateString, FormatAmountPrecision, formattingAmountPrecision } from 'utils'
-import { Settlement as TxSettlement } from 'hooks/useContractBasedVisualizationData'
+import { Settlement as TxSettlement } from 'apps/explorer/components/TransanctionBatchGraph/settlementBuilder'
 import { Network } from 'types'
 import { networkOptions } from 'components/NetworkSelector'
 import ElementsBuilder, { buildGridLayout } from 'apps/explorer/components/TransanctionBatchGraph/elementsBuilder'
